+++ docs/components/UnderConstruction.tsx
@@ -1,0 +1,8 @@
+export default () => {
+  return (
+    <div>
+      <h3>Hey...</h3>
+      <p>This section is under construction, please comeback later.</p>
+    </div>
+  );
+};