+++ docs/pages/About.tsx
@@ -1,0 +1,9 @@
+import UnderConstruction from '../components/UnderConstruction.js';
+
+export default () => {
+  return (
+    <>
+      <UnderConstruction />
+    </>
+  );
+};