+++ docs/uno.config.ts
@@ -1,5 +1,5 @@
 // uno.config.ts
-import { defineConfig } from 'unocss';
+import { defineConfig, presetUno } from 'unocss';
 import { presetWebFonts } from 'unocss';
 
 //  const join = (array: Array<string>): string => `@apply ${array.join(' ')}`;
@@ -13,6 +13,7 @@ export default defineConfig({
         mono: 'Fira Code',
       },
     }),
+    presetUno(),
   ],
   shortcuts: [
     {
@@ -21,9 +22,6 @@ export default defineConfig({
 
       'col-center': 'col justify-center items-center',
       'row-center': 'row justify-center items-center',
-
-      'u-border':
-        'border border-1 border-solid border-[color:var(--border)] hover:border-[color:var(--border-strong)]',
     },
   ],
 });