+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - add `usePromise`, a hook that allows the user to fetch data while tracking the request state and return value.
+- add `joinClasses`, a utility function that filter and returns a valid className as a string.
+- add `getSearchQuery`, a function that returns the search params of the current route as a typed object.
+- add `getRoute`, a function that returns the current url without the `base`.
 
 ## 0.2.0 - 2023-06-02